Module Ocaml_intrinsics

module Crc : sig ... end

module Ext_pointer : sig ... end

Ext_pointer uses values of the OCaml type "int" to represent pointers to 2-byte aligned memory blocks allocated outside the OCaml heap.

module Float : sig ... end

module Int : sig ... end

The are two version of count_leading_zeros, count_set_bits each, which differ in their native code implementation. The first version takes as input a tagged integer and the second version takes as input an untagged integer. Generally, the first version (that operates on a tagged integer) is faster, but if the integer is already untagged, it may be faster to use the second version.

module Int32 : sig ... end

module Int64 : sig ... end

module Native_pointer : sig ... end

Native_pointer uses Nativeint to hold a pointer to a memory block allocated outside the OCaml heap. The pointer is not required to be aligned.

module Nativeint : sig ... end

module Perfmon : sig ... end

module Prefetch : sig ... end

Bring the line of data from memory that contains the specified address.