Module Ext_pointer.Immediate

Parameters

module V : sig ... end

Signature

val unsafe_load_immediate : t -> V.t

load_immediate t assumes without checking that the value pointed to by t is immediate.

val store_immediate : t -> V.t -> unit

store_int t i stores the immediate i to the memory address represented by t.