Module Native_pointer.Bool

val unsafe_load_immediate : t -> Stdlib.Bool.t

load_immediate assumes without checking that the value pointed to by t is an immediate (i.e., the least significant bit is set).

val store_immediate : t -> Stdlib.Bool.t -> unit

store_int t i stores the (tagged) immediate i to the memory address represented by t.