Module Native_pointer.Expert

val of_nativeint : nativeint -> t

These are *not* equivalent to unsafe_of/to_value: they convert between a boxed nativeint and a boxed native pointer (as in Obj.magic), which is always legal because they have the same representation.

val to_nativeint : t -> nativeint