Module Native_pointer.Immediate

Parameters

module V : sig ... end

Signature

val unsafe_load_immediate : t -> V.t

load_immediate assumes without checking that the value pointed to by t is an immediate (i.e., the least significant bit is set).

val store_immediate : t -> V.t -> unit

store_int t i stores the (tagged) immediate i to the memory address represented by t.